rename question actions that are not plain object file
